Model Simplification
I've merged into the Child and Head of Family Models required information that was previously stored outside of the model such as the formatting of the edit table and and for checking the data when making changes.

I've also created a model which the Child and Family Models extend, where shared details and functions related to each instance of a child or family member are stored.

This now needs to be done for all other pages
Then I need to create dummy data for the other pages
Then I need to change the other pages to use the correct dummy data